backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
argon2-cffi==23.1.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
cryptography==42.0.0
pyotp==2.9.0
qrcode[pil]==7.4.2
python-dotenv==1.0.0


